GenoSenseMap/genosense/GenoSenseMap/requirements-prod.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
gunicorn==23.0.0
numpy==1.26.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
email-validator==2.1.1
werkzeug==2.3.7
tensorflow==2.14.0
keras==2.14.0
torch==2.6.0
torchvision==0.21.0
google-cloud-storage==3.1.0
pytest==7.4.0